minor changes to names and moving functions
