+++ src/components/wrappers/ProcessWrapper.tsx
@@ -13,8 +13,9 @@ import { useAppName, useAppOwner } from 'src/core/texts/appTexts';
 import { useCurrentDataModelGuid } from 'src/features/datamodel/useBindingSchema';
 import { FormProvider } from 'src/features/form/FormContext';
 import { useLayoutSets } from 'src/features/form/layoutSets/LayoutSetsProvider';
-import { useCurrentTaskTypeFromProcess, useGetTaskType, useLaxProcessData } from 'src/features/instance/ProcessContext';
+import { useLaxProcessData } from 'src/features/instance/ProcessContext';
 import { ProcessNavigationProvider } from 'src/features/instance/ProcessNavigationContext';
+import { useCurrentTaskTypeFromProcess, useGetTaskType } from 'src/features/instance/useProcessTask';
 import { Lang } from 'src/features/language/Lang';
 import { useLanguage } from 'src/features/language/useLanguage';
 import { PDFWrapper } from 'src/features/pdf/PDFWrapper';
