+++ src/features/instance/ProcessContext.tsx
@@ -6,12 +6,10 @@ import type { UseQueryResult } from '@tanstack/react-query';
 
 import { DisplayError } from 'src/core/errorHandling/DisplayError';
 import { Loader } from 'src/core/loading/Loader';
-import { useApplicationMetadata } from 'src/features/applicationMetadata/ApplicationMetadataProvider';
 import { useLayoutSets } from 'src/features/form/layoutSets/LayoutSetsProvider';
 import { useNavigationParam } from 'src/features/routing/AppRoutingContext';
 import { TaskKeys, useNavigatePage } from 'src/hooks/useNavigatePage';
 import { fetchProcessState } from 'src/queries/queries';
-import { ProcessTaskType } from 'src/types';
 import { behavesLikeDataTask } from 'src/utils/formLayout';
 import type { QueryDefinition } from 'src/core/queries/usePrefetchQuery';
 import type { IProcess } from 'src/types/shared';
@@ -31,30 +29,30 @@ const ProcessContext = createContext<Pick<UseQueryResult<IProcess, HttpClientErr
 );
 
 export function ProcessProvider({ children, instanceId }: PropsWithChildren<{ instanceId: string }>) {
-  const taskId = useNavigationParam('taskId');
+  const taskIdParam = useNavigationParam('taskId');
   const layoutSets = useLayoutSets();
   const navigateToTask = useNavigatePage().navigateToTask;
 
-  const { isLoading, data, error, refetch } = useQuery<IProcess, HttpClientError>(getProcessQueryDef(instanceId));
-
-  useEffect(() => {
-    const elementId = data?.currentTask?.elementId;
-    if (data?.ended) {
-      const hasCustomReceipt = behavesLikeDataTask(TaskKeys.CustomReceipt, layoutSets);
-      if (hasCustomReceipt) {
-        navigateToTask(TaskKeys.CustomReceipt);
-      } else {
-        navigateToTask(TaskKeys.ProcessEnd);
-      }
-    } else if (elementId && elementId !== taskId) {
-      navigateToTask(elementId, { replace: true, runEffect: taskId !== undefined });
-    }
-    //eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [data]);
+  const {
+    isLoading,
+    data: process,
+    error,
+    refetch,
+  } = useQuery<IProcess, HttpClientError>(getProcessQueryDef(instanceId));
 
   useEffect(() => {
     error && window.logError('Fetching process state failed:\n', error);
   }, [error]);
+
+  if (process?.ended) {
+    const hasCustomReceipt = behavesLikeDataTask(TaskKeys.CustomReceipt, layoutSets);
+    hasCustomReceipt ? navigateToTask(TaskKeys.CustomReceipt) : navigateToTask(TaskKeys.ProcessEnd);
+  }
+
+  const currentProcessTaskId = process?.currentTask?.elementId;
+  if (currentProcessTaskId && currentProcessTaskId !== taskIdParam) {
+    navigateToTask(currentProcessTaskId, { replace: true, runEffect: taskIdParam !== undefined });
+  }
 
   if (isLoading) {
     return <Loader reason='fetching-process' />;
@@ -64,90 +62,9 @@ export function ProcessProvider({ children, instanceId }: PropsWithChildren<{ in
     return <DisplayError error={error} />;
   }
 
-  return <ProcessContext.Provider value={{ data, refetch }}>{children}</ProcessContext.Provider>;
+  return <ProcessContext.Provider value={{ data: process, refetch }}>{children}</ProcessContext.Provider>;
 }
 
 export const useHasProcessProvider = () => useContext(ProcessContext) !== undefined;
 export const useLaxProcessData = () => useContext(ProcessContext)?.data;
 export const useReFetchProcessData = () => useContext(ProcessContext)?.refetch;
-
-/**
- * This returns the task type of the current process task, as we got it from the backend
- *
- * @see useCurrentTaskTypeFromProcess
- */
-export function useTaskTypeFromBackend() {
-  const processData = useLaxProcessData();
-
-  if (processData?.ended) {
-    return ProcessTaskType.Archived;
-  }
-
-  if (processData?.currentTask?.altinnTaskType) {
-    return processData.currentTask.altinnTaskType as ProcessTaskType;
-  }
-
-  return ProcessTaskType.Unknown;
-}
-
-/**
- * This returns the task type of the current process task, as we want to use it in the frontend. Some tasks
- * are configured to behave like data tasks, and we want to treat them as such.
- *
- * @see useTaskTypeFromBackend
- */
-export function useCurrentTaskTypeFromProcess() {
-  const taskId = useLaxProcessData()?.currentTask?.elementId;
-  const isStateless = useApplicationMetadata().isStatelessApp;
-  const taskType = useTaskTypeFromBackend();
-  const layoutSets = useLayoutSets();
-
-  if (isStateless) {
-    // Stateless apps only have data tasks. As soon as they start creating an instance from that stateless step,
-    // applicationMetadata.isStatelessApp will return false and we'll proceed as normal.
-    return ProcessTaskType.Data;
-  }
-
-  const isDataTask = behavesLikeDataTask(taskId, layoutSets);
-  return isDataTask ? ProcessTaskType.Data : taskType;
-}
-
-/**
- * This hook returns a function to find the taskType of a given taskId.
- */
-export function useGetTaskType() {
-  const processData = useLaxProcessData();
-  const isStateless = useApplicationMetadata().isStatelessApp;
-  const layoutSets = useLayoutSets();
-
-  return (taskId: string | undefined) => {
-    const task =
-      (processData?.processTasks?.find((t) => t.elementId === taskId) ?? processData?.currentTask?.elementId === taskId)
-        ? processData?.currentTask
-        : undefined;
-
-    if (isStateless) {
-      // Stateless apps only have data tasks. As soon as they start creating an instance from that stateless step,
-      // applicationMetadata.isStatelessApp will return false and we'll proceed as normal.
-      return ProcessTaskType.Data;
-    }
-
-    if (taskId === TaskKeys.CustomReceipt) {
-      return ProcessTaskType.Data;
-    }
-
-    if (taskId === TaskKeys.ProcessEnd) {
-      return ProcessTaskType.Archived;
-    }
-
-    if (processData?.ended) {
-      return ProcessTaskType.Archived;
-    }
-    if (task === undefined || task?.altinnTaskType === undefined) {
-      return ProcessTaskType.Unknown;
-    }
-
-    const isDataTask = behavesLikeDataTask(task.elementId, layoutSets);
-    return isDataTask ? ProcessTaskType.Data : (task.altinnTaskType as ProcessTaskType);
-  };
-}
