+++ src/layout/Button/ButtonComponent.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 
 import { useSetReturnToView } from 'src/features/form/layout/PageNavigationContext';
-import { useLaxProcessData, useTaskTypeFromBackend } from 'src/features/instance/ProcessContext';
+import { useLaxProcessData } from 'src/features/instance/ProcessContext';
 import { useProcessNavigation } from 'src/features/instance/ProcessNavigationContext';
+import { useTaskTypeFromBackend } from 'src/features/instance/useProcessTask';
 import { Lang } from 'src/features/language/Lang';
 import { useLanguage } from 'src/features/language/useLanguage';
 import { getComponentFromMode } from 'src/layout/Button/getComponentFromMode';
